ollama_server/requirements.txt
fastapi
uvicorn
requests
jinja2
python-multipart
sentence-transformers
faiss-cpu
pandas
pymupdf
pyyaml
pydantic